fixes for clippy and file size
